requirements.txt
Flask
reportlab
python-docx
requests